[OGUI-848] Refactor project structre
